Keras-TF: updates to avoid retracing - created a dedicated tf_zoom_in_parameters function with a TF implementation, used in PyramidalInverseCompositional - Ensured in PyramidInverseCompositional that all tensor dimensions are know by autograph, and that p and dp have fixed size with padding - retracings logged into the issue https://github.com/mfournigault/inverse_compositional_algorithm/issues/26 are still there. But they might be due to the fact that for each unit test we create a new instance of PICA with different parameters, and not due to the algorithm itself. This should be tested further with 2 batches of images on which we compute the same transformation type.

+++ src/keras-tf/tf_derivatives.py
@@ -52,8 +52,8 @@ def tf_jacobian(transform_type_value, nx, ny) -> tf.Tensor:
         tf.Tensor: A tensor representing the Jacobian matrix with an added 
         batch dimension.
     """
-    x = tf.range(nx, dtype=tf.float32)
-    y = tf.range(ny, dtype=tf.float32)
+    x = tf.range(tf.cast(nx, dtype=tf.float32), dtype=tf.float32)
+    y = tf.range(tf.cast(ny, dtype=tf.float32), dtype=tf.float32)
     X, Y = tf.meshgrid(x, y)
     ones = tf.ones_like(X)
     zeros = tf.zeros_like(X)
